+++ config.py
@@ -5,6 +5,7 @@
 Configuration script for SM Demo. Configures the GPIOs
 
 Changelogs:
+1.3.0 - removed encoders
 1.2.0 - support for multilanguage
 1.1.0 - volume unit of measure added
 1.0.0 - first release
@@ -16,7 +17,7 @@ __author__ = "Alberto Occelli"
 __copyright__ = "Copyright 2023,"
 __credits__ = ["Alberto Occelli"]
 __license__ = "MIT"
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 __maintainer__ = "Alberto Occelli"
 __email__ = "albertoccelli@gmail.com"
 __status__ = "Dev"
@@ -31,29 +32,17 @@ d_sensor_enabled = load_config(config_file).get("distance_sensor_enabled")  # lo
 # Language
 lang = load_config(config_file).get("lang")  # language
 
-# Buttons and encoders
-# Background volume encoder
-bg_vol_button = load_config(config_file).get("bg_vol_button")  # pause/play button
-bg_vol_dt_pin = load_config(config_file).get("bg_vol_dt_pin")  # rotary encoder DT pin
-bg_vol_clk_pin = load_config(config_file).get("bg_vol_clk_pin")  # rotary encoder CLK pin
-# Background track encoder
-bg_tr_button = load_config(config_file).get("bg_tr_button")  # pause/play button
-bg_tr_dt_pin = load_config(config_file).get("bg_tr_dt_pin")  # rotary encoder DT pin
-bg_tr_clk_pin = load_config(config_file).get("bg_tr_clk_pin")  # rotary encoder CLK pin
-# Front volume encoder
-fr_vol_button = load_config(config_file).get("fr_vol_button")  # pause/play button
-fr_vol_dt_pin = load_config(config_file).get("fr_vol_dt_pin")  # rotary encoder DT pin
-fr_vol_clk_pin = load_config(config_file).get("fr_vol_clk_pin")  # rotary encoder CLK pin
-# Front track encoder
-fr_tr_button = load_config(config_file).get("fr_tr_button")  # pause/play button
-fr_tr_dt_pin = load_config(config_file).get("fr_tr_dt_pin")  # rotary encoder DT pin
-fr_tr_clk_pin = load_config(config_file).get("fr_tr_clk_pin")  # rotary encoder CLK pin
+# Buttons
+# Front track button
+fr_tr_button = load_config(config_file).get("fr_tr_button")
+# Background track button
+bg_tr_button = load_config(config_file).get("bg_tr_button")
+# Front volume up button
+fr_vol_up_button = load_config(config_file).get("fr_vol_up_button")
+# Front volume down button
+fr_vol_down_button = load_config(config_file).get("fr_vol_up_button")
 
-# Sensor
-echo_pin = load_config(config_file).get("echo_pin")  # sensor echo pin
-trig_pin = load_config(config_file).get("trig_pin")  # sensor trigger pin
 
-# Player settings
 start_track = load_config(config_file).get("start_track") - 1  # start track
 vol_step_um = load_config(config_file).get("volume_steps_um")  # unit of measure for volume steps
 vol_step = load_config(config_file).get("volume_steps")  # volume steps
@@ -63,15 +52,7 @@ lang = load_config(config_file).get("lang")  # language
 
 # Setup GPIOs as inputs
 GPIO.setmode(GPIO.BCM)
-GPIO.setup(bg_vol_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(bg_vol_dt_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(bg_vol_clk_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(bg_tr_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(bg_tr_dt_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(bg_tr_clk_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(fr_vol_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(fr_vol_dt_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(fr_vol_clk_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 GPIO.setup(fr_tr_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(fr_tr_dt_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(fr_tr_clk_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(bg_tr_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(fr_vol_up_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+GPIO.setup(fr_vol_down_button, GPIO.IN, pull_up_down=GPIO.PUD_UP)